Rename record's template directory to hbs
