+++ packages/record/hbs/consent-video-trial.hbs
