+++ packages/record/hbs/video-config.hbs
