+++ packages/record/src/recorder.ts
@@ -3,11 +3,11 @@ import LookitS3 from "@lookit/data/dist/lookitS3";
 import autoBind from "auto-bind";
 import Handlebars from "handlebars";
 import { JsPsych } from "jspsych";
+import playbackFeed from "../hbs/playback-feed.hbs";
+import recordFeed from "../hbs/record-feed.hbs";
+import webcamFeed from "../hbs/webcam-feed.hbs";
 import play_icon from "../img/play-icon.svg";
 import record_icon from "../img/record-icon.svg";
-import playbackFeed from "../templates/playback-feed.hbs";
-import recordFeed from "../templates/record-feed.hbs";
-import webcamFeed from "../templates/webcam-feed.hbs";
 import {
   CreateURLError,
   NoStopPromiseError,
