+++ packages/record/src/stop.ts
@@ -1,7 +1,7 @@
 import { LookitWindow } from "@lookit/data/dist/types";
 import Handlebars from "handlebars";
 import { JsPsych, JsPsychPlugin } from "jspsych";
-import uploadingVideo from "../templates/uploading-video.hbs";
+import uploadingVideo from "../hbs/uploading-video.hbs";
 import { NoSessionRecordingError } from "./errors";
 import Recorder from "./recorder";
